Moved JWT signature to oidc-client-dpop-browser
The dependency on crypto libraries used to sign the JWT that is used as a DPoP header has been pushed down one level.
However, this created some issues with rollup. For the time being, since the library is not exposed to be used directly in the browser,
the bundler isn't used anymore, and the library is just transpiled before being consumed by the browser module.
