+++ packages/browser/__tests__/jose/IsomorphicJoseUtility.test.ts
@@ -20,30 +20,7 @@
  */
 
 import { describe, it } from "@jest/globals";
-import {
-  decodeJWT,
-  generateCodeVerifier,
-  signJWT,
-} from "../../src/jose/IsomorphicJoseUtility";
-import { generateKeyForDpop } from "@inrupt/oidc-dpop-client-browser";
-
-describe("signJWT/decodeJWT", () => {
-  it("generates a JWT that can be decoded without signature verification", async () => {
-    const key = await generateKeyForDpop();
-    const payload = { testClaim: "testValue" };
-    const jwt = await signJWT(payload, key, { algorithm: "RS256" });
-    const decoded = await decodeJWT(jwt);
-    expect(decoded.testClaim).toEqual(payload.testClaim);
-  });
-
-  it("can verify the ES256 signature of the generated JWT", async () => {
-    const key = await generateKeyForDpop();
-    const payload = { testClaim: "testValue" };
-    const jwt = await signJWT(payload, key, { algorithm: "ES256" });
-    const decoded = await decodeJWT(jwt, key, { algorithms: ["ES256"] });
-    expect(decoded.testClaim).toEqual(payload.testClaim);
-  });
-});
+import { generateCodeVerifier } from "../../src/jose/IsomorphicJoseUtility";
 
 describe("generateCodeVerifier", () => {
   it("generates a different string on each call", async () => {
