+++ packages/browser/src/jose/IsomorphicJoseUtility.ts
@@ -28,21 +28,22 @@
  * File for NodeJS-compatible JOSE
  */
 import {
-  JWKECKey,
   ECCurve,
   BasicParameters,
   OKPCurve,
-  JWKOKPKey,
-  JWKRSAKey,
-  JWKOctKey,
-  JWT as JoseJWT,
   JSONWebKey,
+  // JWKECKey,
+  // JWKOKPKey,
+  // JWKRSAKey,
+  // JWKOctKey,
+  // JWT as JoseJWT,
 } from "jose";
 import { JWK } from "node-jose";
-import JWT, { VerifyOptions } from "jsonwebtoken";
 import IJoseUtility from "./IJoseUtility";
 import randomString from "crypto-random-string";
 import crypto from "crypto";
+import { signJWT, decodeJWT } from "@inrupt/oidc-dpop-client-browser";
+// import JWT, { VerifyOptions } from "jsonwebtoken";
 
 /**
  * Generates a Json Web Key
@@ -57,52 +58,6 @@ export async function generateJWK(
 ): Promise<JSONWebKey> {
   const key = await JWK.createKey(kty, crvBitlength, parameters);
   return key.toJSON(true) as JSONWebKey;
-}
-
-/**
- * Generates a Json Web Token (https://tools.ietf.org/html/rfc7519) containing
- * the provided payload and using the signature algorithm specified in the options.
- * @param payload The body of the JWT.
- * @param key The key used for the signature.
- * @param options
- * @returns a 3-parts base64-encoded string, split by dots.
- * @hidden
- */
-export async function signJWT(
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  payload: Record<string, any>,
-  key: JWKECKey | JWKOKPKey | JWKRSAKey | JWKOctKey,
-  options?: JoseJWT.SignOptions
-): Promise<string> {
-  const parsedKey = await JWK.asKey(key);
-  const convertedKey: string = parsedKey.toPEM(true);
-  const signed = JWT.sign(payload, convertedKey, {
-    ...(options as JWT.SignOptions),
-  });
-  return signed;
-}
-
-/**
- * Decodes the base64 Json Web Token into an object. If a key is specified, the
- * JWT is also verified.
- * @param token The base64-encoded token
- * @param key The key used to sign the token
- * @returns the payload of the JWT
- * @hidden
- */
-export async function decodeJWT(
-  token: string,
-  key?: JWKECKey | JWKOKPKey | JWKRSAKey | JWKOctKey,
-  options?: VerifyOptions
-): Promise<Record<string, unknown>> {
-  if (key) {
-    const parsedKey = await JWK.asKey(key);
-    const convertedKey: string = parsedKey.toPEM(false);
-    return JWT.verify(token, convertedKey, options) as Promise<
-      Record<string, unknown>
-    >;
-  }
-  return JWT.decode(token) as Promise<Record<string, unknown>>;
 }
 
 /**
