+++ packages/oidc-dpop-client-browser/package.json
@@ -13,7 +13,8 @@
     "clean": "npm run clean-module && npm run clean-browser && npm run clean-examples && rm -rf ./coverage",
     "clean-module": "rm -rf ./dist",
     "clean-browser": "rm -rf ./browserDist",
-    "build": "rollup --config rollup.config.js",
+    "rollup": "rollup --config rollup.config.js",
+    "build": "tsc -p tsconfig.json",
     "lint": "eslint --max-warnings=0 \"src/**/*.ts\"",
     "lint-fix": "eslint --max-warnings=0 --fix \"src/**\"",
     "format": "prettier --write \"src/**\" \"__tests__/**\"",
@@ -29,15 +30,16 @@
     "@rollup/plugin-node-resolve": "^9.0.0",
     "cross-fetch": "^3.0.6",
     "rollup": "^2.15.0",
-    "rollup-plugin-node-builtins": "^2.1.2",
-    "rollup-plugin-node-polyfills": "^0.2.1",
-    "rollup-plugin-typescript2": "^0.27.1"
+    "rollup-plugin-typescript2": "^0.27.1",
+    "rollup-plugin-node-polyfills": "^0.2.1"
   },
   "dependencies": {
     "@inrupt/solid-client-authn-core": "^0.2.0",
+    "@types/jsonwebtoken": "^8.5.0",
     "@types/node-jose": "^1.1.5",
     "@types/url-parse": "^1.4.3",
     "jose": "^2.0.2",
+    "jsonwebtoken": "^8.5.1",
     "node-jose": "^2.0.0",
     "oidc-client": "^1.10.1",
     "url-parse": "^1.4.7"
